% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{delete}
\alias{delete}
\title{Deleta a Value or Directory from an Object Store
Delete the value or directory corresponding to the specified \code{key} from the object store service
connection \code{con}.}
\usage{
delete(con, key)
}
\arguments{
\item{con}{An object store connection from \code{\link{register_service}}.}

\item{key}{A key name, optionally including a \code{/} separated directory path}
}
\value{
\code{NULL} is invisibly returned, or an error may be thrown.
}
\description{
Deleta a Value or Directory from an Object Store
Delete the value or directory corresponding to the specified \code{key} from the object store service
connection \code{con}.
}
\seealso{
\code{\link{register_service}} \code{\link{cache}} \code{\link{uncache}}
}

