% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{htdigest}
\alias{htdigest}
\title{Create or Update a Digest Authentication File}
\usage{
htdigest(file, realm, user, password)
}
\arguments{
\item{file}{An 'htdigest'-style file to create or update.}

\item{realm}{The authentication realm.}

\item{user}{User name.}

\item{password}{User password.}
}
\value{
Invoked for the side effect of creating or updating \code{file}.
}
\description{
Use this function to create or edit 'htdigest'-style HTTP digest authentication
files for the \code{\link{mongoose}} back end. Depending on the authentication
settings (global or per-directory), you may use just one global file or a file
in each subdirectory of the mongoose data directory.
}
\note{
See \url{https://httpd.apache.org/docs/current/programs/htdigest.html}.
}

