% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{register_service}
\alias{register_service}
\title{Register an Object Store Service}
\usage{
register_service(url = "http://localhost:8000", backend = mongoose, ...)
}
\arguments{
\item{url}{The service root url including protocol, address and port number. For example \code{http://localhost:8000}.}

\item{backend}{A service backend provider. The default is \code{mongoose}, but you may choose from other available
backends like \code{minio} and {s3}.}

\item{...}{Backend-specific arguments, see backend documentation for details.}
}
\value{
A function used by \code{\link{cache}}, \code{\link{uncache}} and \code{\link{delete}} to access the service.
}
\description{
Register an object store service backend, including backend-specific options
like authentication, encryption, and compression.
}
\examples{
# Start an example local mongoose backend server
mongoose_start()
con <- register_service()
# Cache the 'iris' dataset in a directory named 'mydata':
cache(con, iris, "mydata/iris")
# Retrieve it from the cache into a new variable called 'x'
x <- uncache(con, "mydata/iris")
# Delete the entire 'mydata' directory
delete(con, "mydata")
mongoose_stop()
}
\seealso{
\code{\link{mongoose}} \code{link{cache}} \code{\link{uncache}} \code{\link{delete}}
}

