% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{uncache}
\alias{uncache}
\title{Retrieve a Value from an Object Store}
\usage{
uncache(con, key = "")
}
\arguments{
\item{con}{An object store connection from \code{\link{register_service}}.}

\item{key}{A key name, optionally including a \code{/} separated directory path.}
}
\value{
Either a data frame directory listing when \code{key} corresponds to a directory,
or an R value corresponding to \code{key}.
}
\description{
Retrieve a value corresponding to the specified \code{key} from the object store service
connection \code{con}. If \code{key} corresponds to a directory path, then a data frame listing
the directory contents is returned. Set \code{key=""} to list the contents of the
service root directory path.
}
\note{
Directory entries in the data frame directory listing output are identified by \code{size=NA}.
}
\examples{
# Start an example local mongoose backend server
mongoose_start()
con <- register_service()
# Cache the 'iris' dataset in a directory named 'mydata':
cache(con, iris, "mydata/iris")
# Retrieve it from the cache into a new variable called 'x'
x <- uncache(con, "mydata/iris")
# Delete the entire 'mydata' directory
delete(con, "mydata")
mongoose_stop()
}
\seealso{
\code{\link{register_service}} \code{\link{cache}} \code{\link{delete}}
}

