% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{delete}
\alias{delete}
\title{Deleta a Value or Directory from an Object Store}
\usage{
delete(con, key)
}
\arguments{
\item{con}{An object store connection from \code{\link{register_service}}.}

\item{key}{A key name, optionally including a \code{/} separated directory path}
}
\value{
\code{NULL} is invisibly returned, or an error may be thrown.
}
\description{
Delete the value or directory corresponding to the specified \code{key} from the object store service
connection \code{con}.
}
\examples{
# Start an example local mongoose backend server
mongoose_start()
con <- register_service()
# Cache the 'iris' dataset in a directory named 'mydata':
cache(con, iris, "mydata/iris")
# Retrieve it from the cache into a new variable called 'x'
x <- uncache(con, "mydata/iris")
# Delete the entire 'mydata' directory
delete(con, "mydata")
mongoose_stop()
}
\seealso{
\code{\link{register_service}} \code{\link{cache}} \code{\link{uncache}}
}

