% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongoose.R
\name{mongoose}
\alias{mongoose}
\title{Mongoose Object Store Backend
Specify connection details for a \code{mongoose} web service object store backend,
the default simple back end included in the \code{feathercache} package.}
\usage{
mongoose(uri, ...)
}
\arguments{
\item{uri}{The serivce uri, for instance \code{http://localhost:8000}.}

\item{...}{Optional service parameters including:
\itemize{
  \item{user}{Optional HTTP digest authentication user name}
  \item{password}{Optional HTTP digest authentication user password}
  \item{ssl_verifyhost}{Optional SSL/TLS host verification, defaults to 0 (no verification)}
  \item{ssl_verifypeer}{Optional SSL/TLS peer verification, defaults to 0 (no verification)}
}}
}
\description{
Mongoose Object Store Backend
Specify connection details for a \code{mongoose} web service object store backend,
the default simple back end included in the \code{feathercache} package.
}

