% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongoose.R
\name{mongoose_start}
\alias{mongoose_start}
\title{Start a Mongoose Service
Manuall start a local mongoose service}
\usage{
mongoose_start(port = 8000L, directory = getwd(), forward_to = NULL,
  ssl_cert = NULL, auth_domain = NULL, global_auth = NULL,
  directory_auth = NULL)
}
\arguments{
\item{port}{service port number}

\item{directory}{full path to data directory}

\item{forward_to}{forward 'not found' requests to another server}

\item{ssl_cert}{TLS/SSL certificate}

\item{auth_domain}{HTTP digest authentication domain/realm}

\item{global_auth}{HTTP digest global authentication file}

\item{directory_auth}{HTTP digest per-directory authentication file name}
}
\value{
Nothing; the mongoose server is started up as a background process.
}
\description{
Start a Mongoose Service
Manuall start a local mongoose service
}
\note{
Leave parameters \code{NULL} to not use the corresponding features.
}

